Rename PackageDataFetcher to calls
Rename the api/PackageDataFetcher module to api/calls, as it's likely to
be expanded later on with other API call logic.

Refs TS-1074
